services/comm-service/requirements.txt
# Django
Django==5.0.1
django-cors-headers==4.3.1

# API Frameworks
django-ninja==1.1.0
djangorestframework==3.14.0

# Real-time WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Object Storage
minio==7.2.0
boto3==1.34.14

# Database
psycopg2-binary==2.9.9

# Server
gunicorn==21.2.0
uvicorn[standard]==0.25.0

# Consul integration
python-consul==1.1.0
requests==2.31.0

# Metrics (Prometheus)
django-prometheus==2.3.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pika==1.3.2
